removed some unneeded code from client side, updated readme files, and I extracted the job search related things from app.js to put it in it's own component
